requirements.txt
Pillow==10.1.0
pdf2image==1.16.3
img2pdf==0.5.1
python-docx==1.1.0
docx2pdf==0.1.8
pypdf2==3.0.1
pdf2docx==0.5.8
